requirements.txt
digi-xbee==1.5.0
paho-mqtt==2.1.0
pyserial==3.5
